+++ tests/test_report.py
@@ -4,10 +4,25 @@
 
 """Tests for reuse.report"""
 
+# pylint: disable=invalid-name
+
 import os
+import sys
+
+import pytest
 
 from reuse.project import Project
 from reuse.report import FileReport, ProjectReport
+
+try:
+    import posix as is_posix
+except ImportError:
+    is_posix = False
+
+cpython = pytest.mark.skipif(
+    sys.implementation.name != "cpython", reason="only CPython supported"
+)
+posix = pytest.mark.skipif(not is_posix, reason="Windows not supported")
 
 
 def test_generate_file_report_file_simple(fake_repository):
@@ -117,7 +132,7 @@ def test_generate_project_report_missing_license(
 
 def test_generate_project_report_bad_license(fake_repository, multiprocessing):
     """Bad licenses are detected."""
-    (fake_repository / "LICENSES/bad.txt").touch()
+    (fake_repository / "LICENSES/bad.txt").write_text("foo")
 
     project = Project(fake_repository)
     result = ProjectReport.generate(project, multiprocessing=multiprocessing)
@@ -165,9 +180,12 @@ def test_generate_project_report_deprecated_license(
     assert "GPL-3.0" in result.deprecated_licenses
 
 
+@cpython
+@posix
 def test_generate_project_report_read_error(fake_repository, multiprocessing):
     """Files that cannot be read are added to the read error list."""
-    (fake_repository / "bad").symlink_to("does_not_exist")
+    (fake_repository / "bad").write_text("foo")
+    (fake_repository / "bad").chmod(0o000)
 
     project = Project(fake_repository)
     result = ProjectReport.generate(project, multiprocessing=multiprocessing)